refactor(web): move technology status helper to utils
